- Fixed a few warnings in TextArea. - Discovered a bug in Population that would only clear half a list when clear() was called. - Added a case where if there's no food all colonists die and the game ends.
git-svn-id: https://svn.outpostuniverse.org:8443/svn/outposthd@221 bdf9137d-f3be-7946-88fe-7b0bbfdbf31a

+++ src/UI/Core/TextArea.cpp
@@ -36,7 +36,7 @@ void TextArea::processString()
 
 	StringList tokenList = split(text().c_str());
 	
-	int w = 0, i = 0;
+	size_t w = 0, i = 0;
 	while (i < tokenList.size())
 	{
 		string line;
@@ -70,7 +70,7 @@ void TextArea::processString()
 		mFormattedList.push_back(line);
 	}
 
-	mNumLines = height() / font().height();
+	mNumLines = static_cast<size_t>(height() / font().height());
 }
